Clean up main class Move listener to own class for more OOP approach
